pyproject.toml
[tool.poetry]
name = "library-django"
version = "0.1.0"
description = ""
authors = ["Pavel2232 <106510078+Pavel2232@users.noreply.github.com>"]
readme = "README.md"
packages = [{include = "library_django"}]

[tool.poetry.dependencies]
python = "^3.11"
django = "^4.2.5"
django-environ = "^0.11.2"
psycopg2 = "^2.9.7"
djangorestframework = "^3.14.0"
drf-spectacular = "^0.26.5"
django-filter = "^23.3"
gunicorn = "^21.2.0"


[tool.poetry.group.dev.dependencies]
pytest-django = "^4.5.2"
faker = "^19.6.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
